Refactor commit process with modular architecture
